requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
array_api_compat==1.10.0
async-timeout==4.0.3
attrs==24.3.0
bcrypt==4.2.1
boto3==1.35.89
botocore==1.35.89
bracex==2.5.post1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.4
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dargs==0.4.10
dataclasses-json==0.6.7
deepmd-kit==3.0.1
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
emmet-core==0.84.6rc2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastavro==1.10.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.13
langchain-cohere==0.3.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
latexcodec==3.0.0
maggma==0.70.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
mendeleev==0.19.0
ml_dtypes==0.5.0
mongomock==4.3.0
monty==2024.12.10
mp-api==0.45.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.4
openai==1.58.1
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
palettable==3.3.3
pandas==2.2.3
parameterized==0.9.0
paramiko==3.5.0
pillow==11.0.0
plotly==5.24.1
propcache==0.2.1
protobuf==5.29.2
pybtex==0.24.0
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydash==8.0.4
pyfiglet==0.8.post1
Pygments==2.18.0
pymatgen==2024.11.13
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
scipy==1.14.1
sentinels==1.0.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spglib==2.5.0
SQLAlchemy==2.0.36
sshtunnel==0.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.0
tqdm==4.67.1
triton==3.1.0
typeguard==4.4.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uncertainties==3.2.2
urllib3==2.3.0
wcmatch==10.0
wrapt==1.17.0
yarl==1.18.3
